Backend/requirements-fastapi.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
google-genai>=1.13.0
